refactor handle.go to internal
